requirements.txt
dash==3.0.3
dash_bootstrap_components==2.0.2
dash_bootstrap_templates==2.1.0
pandas==2.2.3
Pillow==11.2.1
plotly==6.0.1
python-dotenv==1.1.0
tensorflow==2.15.0
gunicorn==23.0.0